requirements.txt
asgiref==3.2.7
backcall==0.1.0
colorama==0.4.3
decorator==4.4.2
Django==3.0.6
django-cors-headers==3.3.0
django-filter==2.2.0
django-guardian==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
ipython==7.15.0
ipython-genutils==0.2.0
jedi==0.17.0
parso==0.7.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
psycopg2==2.8.5
Pygments==2.6.1
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
wcwidth==0.2.2
